+++ ImTK/UI/Element/Basic/Window.cs
@@ -102,6 +102,8 @@ namespace ImTK.UI
         
         public string windowId { get; init; }
 
+        public virtual bool hideInHierarchy => false;
+
         public override string persistenceKey 
         {
             get