+++ ImTK/UI/IRenderRoot.cs
@@ -2,5 +2,6 @@ namespace ImTK.UI
 {
     public interface IRenderRoot
     {
+        bool hideInHierarchy { get; }
     }
 }